#48. VideoVault - Video Sharing Platform/backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework-simplejwt==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-redis==5.4.0
django-storages==1.14.4
django-cleanup==9.0.0
django-extensions==3.2.3

psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
gunicorn==23.0.0

Pillow==11.1.0
python-magic==0.4.27
python-decouple==3.8
whitenoise==6.8.2
drf-spectacular==0.28.0

numpy==2.2.1
scikit-learn==1.6.1
